+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## 2.0.27
 
-- Updated `spawnDockerCompose` so it doesn't throw if error code 130 is generated from "ctrl c"
+- Updated `spawnDockerCompose` so it doesn't throw an error if code 130 is generated from process termination via "ctrl + c" 
 
 ## 2.0.28
 
@@ -10,8 +10,12 @@
 
 ## 2.0.29
 
-- Added the util method `splitByWhitespace`
+- Added `splitByWhitespace`method to general utils
 
 ## 2.0.30
 
-- Added `whichWsl` method
+- Added `whichWsl` method to general utils
+
+## 2.0.31
+
+- Added `dockerComposeBash` method to docker utils